+++ spotipy/oauth2.py
@@ -16,7 +16,6 @@ import sys
 import time
 
 import requests
-
 # Workaround to support both python 2 & 3
 import six
 import six.moves.urllib.parse as urllibparse
@@ -135,32 +134,41 @@ class SpotifyOAuth(object):
         self.cache_path = cache_path
         self.scope = self._normalize_scope(scope)
         self.proxies = proxies
+        self.token_info = None
 
     def get_cached_token(self):
         ''' Gets a cached auth token
         '''
-        token_info = None
+        if self.token_info is not None:
+            self._refresh_token_if_expired()
+            return self.token_info
+
         if self.cache_path:
             try:
                 f = open(self.cache_path)
                 token_info_string = f.read()
                 f.close()
-                token_info = json.loads(token_info_string)
+                self.token_info = json.loads(token_info_string)
 
                 # if scopes don't match, then bail
-                if 'scope' not in token_info or not self._is_scope_subset(
-                        self.scope, token_info['scope']):
+                if 'scope' not in self.token_info or not self._is_scope_subset(
+                        self.scope, self.token_info['scope']):
                     return None
 
-                if self.is_token_expired(token_info):
-                    token_info = self.refresh_access_token(
-                        token_info['refresh_token'])
+                self._refresh_token_if_expired()
 
             except IOError:
                 pass
-        return token_info
+
+        return self.token_info
+
+    def _refresh_token_if_expired(self):
+        if self.is_token_expired(self.token_info):
+            self.token_info = self.refresh_access_token(
+                self.token_info['refresh_token'])
 
     def _save_token_info(self, token_info):
+        self.token_info = token_info
         if self.cache_path:
             try:
                 f = open(self.cache_path, 'w')