+++ spotipy/util.py
@@ -103,9 +103,7 @@ def prompt_for_user_token(username, scope=None, client_id=None,
         print()
 
         code = sp_oauth.parse_response_code(response)
-        token_info = sp_oauth.get_access_token(code)
+        sp_oauth.get_access_token(code)
+
     # Auth'ed API request
-    if token_info:
-        return token_info['access_token']
-    else:
-        return None
+    return sp_oauth